Move eventlog to window.Window object
